Alienshooter.py code
This is the rest of the code. Obviously we can’t test this until we get
the graphics going, but I think I hit everything I could. I probably
missed a few things here and there but for the most part, I think we
just need to adjust the code when we put the graphics in